% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cgm.R
\name{CgmObject}
\alias{CgmObject}
\title{Make New CGM Object}
\usage{
CgmObject(con)
}
\arguments{
\item{con}{valid database connection}
}
\description{
The new object contains all data relevant to CGM operations:  `glucose_records`, `notes_records` and more.
}
\details{
CGM-related data structures and functions
}
