% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_experiments_list}
\alias{db_experiments_list}
\title{List all valid experiments}
\usage{
db_experiments_list(con)
}
\arguments{
\item{con}{valid database connection}
}
\value{
character vector of all experiment names
}
\description{
Return a vector of all experiments currently available on the platform.
Unlike [taster::db_food_list()], it uses the values from the database table `experiments`.
so it will not find experiments (usually food names) for some people. On the other hand,
this calls the database each time, so you'll see performance problems if you do it too often.
}
