% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_food_list}
\alias{db_food_list}
\title{List all products consumed by `user_id`}
\usage{
db_food_list(con, user_id = 1234)
}
\arguments{
\item{con}{valid database connection}

\item{user_id}{vector of user IDs or NULL to show all users}
}
\value{
character vector of product names sorted alphabetically
}
\description{
Return all products consumed by this user. If `user_id==NULL` then show
 from all users, all products consumed more than once.
}
