% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_get_table}
\alias{db_get_table}
\title{Get a table from the database and return it as a well-formed dataframe.}
\usage{
db_get_table(con, table_name = "glucose_records")
}
\arguments{
\item{con}{valid database connection}

\item{table_name}{string name for a table in the database (usual `glucose_records` or `notes_records`)}
}
\description{
As much as possible, route calls to the database table to here. This function will
abstract away problems with differences between databases.

hardcoded to look for Tastermonial db
}
