% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_insert_user}
\alias{db_insert_user}
\title{Insert a new `user_id` into appropriate tables}
\usage{
db_insert_user(con, user)
}
\arguments{
\item{con}{valid database connection}

\item{user}{list containing `$user_id` and `$firebase_id`}
}
\value{
integer total of rows affected in database write operations. You have a problem if this is 0.
}
\description{
Insert a new `user_id` into appropriate tables
}
