% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_user_all_ids}
\alias{db_user_all_ids}
\title{All user ids in the database}
\usage{
db_user_all_ids(con)
}
\arguments{
\item{conn_args}{database connection}
}
\value{
vector of all known `user_id`
}
\description{
All user ids in the database
}
