% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_user_df}
\alias{db_user_df}
\title{All user records in the database}
\usage{
db_user_df(conn_args = config::get("dataconnection"))
}
\arguments{
\item{conn_args}{database connection}
}
\value{
dataframe of all user records
}
\description{
All user records in the database
}
