% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_user_id_from_firebase}
\alias{db_user_id_from_firebase}
\title{Find User From Firebase ID}
\usage{
db_user_id_from_firebase(firebase_id)
}
\arguments{
\item{firebase_id}{valid firebase ID for a signed in user}
}
\value{
numeric user_id; NA if no user exists or if the table is not in the database
}
\description{
A signed-in user, by definition, has a Firebase ID.  Because the ID was
registered as a unique `user_id` when the account was created, we can look it up in
the user table and return its user (and username)
}
