% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_user_privileges}
\alias{db_user_privileges}
\title{Returns privileges assigned to this user}
\usage{
db_user_privileges(con, user_id = -1)
}
\arguments{
\item{con}{valid database connection}

\item{user_id}{user ID}
}
\value{
privilege level, NULL if no privileges available
}
\description{
Returns privileges assigned to this user
}
