% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_users_visible}
\alias{db_users_visible}
\title{Users Visible at Privilege}
\usage{
db_users_visible(user_id = -1)
}
\arguments{
\item{user_id}{user ID}
}
\value{
vector of user lists in the form (`user_id`, `username`)
}
\description{
Users Visible at Privilege
}
