% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_write_notes_table}
\alias{db_write_notes_table}
\title{Write a notes table to the database}
\usage{
db_write_notes_table(
  con = db_connection(),
  table_name = "raw_notes",
  table_df,
  user_id
)
}
\arguments{
\item{con}{valid database connection}

\item{table_name}{char string for table name}

\item{table_df}{valid dataframe to write to the database}

\item{user_id}{User ID}
}
\value{
character string with a message that can be displayed to the user
}
\description{
Write a notes table to the database
}
