% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{db_write_table}
\alias{db_write_table}
\title{Write a table to the database}
\usage{
db_write_table(con = db_connection(), table_name = "raw_glucose", table_df)
}
\arguments{
\item{con}{valid database connection}

\item{table_name}{char string for table name}

\item{table_df}{valid dataframe to write to the database}
}
\description{
Write a table to the database
}
