% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{firebase_setup}
\alias{firebase_setup}
\title{Set up Firebase support}
\usage{
firebase_setup(con)
}
\arguments{
\item{con}{database connection}
}
\value{
firebase object
}
\description{
This is just a stub for now, in order to consolidate all program-wide calls to Firebase.
See configuration instructions \url{https://firebase.john-coene.com/guide/config/}
}
