% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_goddess.R
\name{food_df_}
\alias{food_df_}
\title{Food Times Dataframe}
\usage{
food_df_(
  cgm_data,
  user_id,
  timewindow,
  prefixLength,
  food_name,
  normalize = FALSE
)
}
\arguments{
\item{cgm_data}{cgmObject}

\item{user_id}{user ID}

\item{timewindow}{integer minutes to show after start}

\item{prefixLength}{integer minutes before start}

\item{food_name}{character string for food name}

\item{normalize}{logical if output should be normalized}
}
\value{
dataframe of food times
}
\description{
Food Times Dataframe
}
