% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_goddess.R
\name{food_selection}
\alias{food_selection}
\title{Make a Food Selection Object}
\usage{
food_selection(item_id, prod_list, user_id)
}
\arguments{
\item{item_id}{character string NS id for UI}

\item{prod_list}{list of products to choose from}

\item{user_id}{user ID}
}
\value{
food selection object
}
\description{
Make a Food Selection Object
}
