% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{get_id}
\alias{get_id}
\title{#' @describeIn User Getter for user ID
setMethod("get_id", User, function(User) {
  User@.Data[["id"]]
})}
\usage{
get_id(x, ...)
}
\description{
#' @describeIn User Getter for user ID
setMethod("get_id", User, function(User) {
  User@.Data[["id"]]
})
}
\details{
return the value for `id`
}
