% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cgm.R
\name{glucose_ranges_for_id}
\alias{glucose_ranges_for_id}
\title{Max and Min Glucose Ranges for `user_id`}
\usage{
glucose_ranges_for_id(user_id, glucose_records)
}
\arguments{
\item{user_id}{user ID}

\item{glucose_records}{valid glucose records}
}
\description{
Max and Min Glucose Ranges for `user_id`
}
