% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_goddess.R
\name{goddessPlot}
\alias{goddessPlot}
\title{Return a CGM Goddess Plot}
\usage{
goddessPlot(
  user_id,
  food_times_df,
  food_name,
  cgm_data,
  normalize = FALSE,
  smooth = FALSE,
  baseline = FALSE
)
}
\arguments{
\item{user_id}{ID}

\item{food_name}{character string for food name}

\item{cgm_data}{CgmObject}

\item{food_df}{dataframe of food times}
}
\value{
ggplot2 object
}
\description{
Return a CGM Goddess Plot
}
