% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{name_for_user_id}
\alias{name_for_user_id}
\title{Character string for user_id}
\usage{
name_for_user_id(con, firebase_obj, user_id)
}
\arguments{
\item{con}{database connection}

\item{firebase_obj}{firebase object}

\item{user_id}{numerical `user_id`}
}
\value{
dataframe of all user records
}
\description{
Character string for user_id
}
