% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_compare_glucose}
\alias{plot_compare_glucose}
\title{Plot a glucose dataframe comparing several meals}
\usage{
plot_compare_glucose(
  food_times_df,
  combine = FALSE,
  smooth = FALSE,
  title = "Name",
  subtitle = "foodname",
  legend_var = "date_ch"
)
}
\arguments{
\item{food_times_df}{dataframe of a valid dataframe in food_times_df format.}

\item{combine}{combine lines into one average (default = FALSE)}

\item{smooth}{smooth lines (default = FALSE)}

\item{title}{string to display on ggplot}

\item{subtitle}{string to display for subtitle}

\item{legend_var}{a food_times_df column name to be used for the plot legend}
}
\value{
ggplot object
}
\description{
Plot of a valid CGM file.
}
