% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plot_glucose}
\alias{plot_glucose}
\title{Plot a glucose dataframe}
\usage{
plot_glucose(glucose_raw, title = "Name", subtitle = "")
}
\arguments{
\item{glucose_raw}{dataframe of a valid CGM data stream}

\item{title}{string to display on ggplot}

\item{subtitle}{string to display on ggplot}
}
\value{
ggplot object
}
\description{
Plot of a valid CGM file.
}
