% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{print_user}
\alias{print_user}
\title{Print this user}
\usage{
print_user(userObj)
}
\arguments{
\item{userObject}{user object}
}
\value{
character string of user object
}
\description{
Print this user
}
