% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\docType{class}
\name{RegisteredUsers-class}
\alias{RegisteredUsers-class}
\alias{RegisteredUsers}
\alias{initialize,RegisteredUsers-method}
\title{Registered Users From Database}
\usage{
\S4method{initialize}{RegisteredUsers}(.Object, con = NULL, firebase_obj = NULL)
}
\arguments{
\item{con}{database connection (optional: if missing, call \code{db_connection})}

\item{firebase_obj}{a valid Firebase object}
}
\description{
Registered Users From Database
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: initialize
}}

\section{Slots}{

\describe{
\item{\code{con}}{valid database connection}

\item{\code{users_df}}{dataframe of users}

\item{\code{glucose_records}}{dataframe of glucose}

\item{\code{notes_records}}{dataframe of notes}
}}

