% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{scaled_axis}
\alias{scaled_axis}
\title{Adjust x axis depending on time scale of glucose data frame}
\usage{
scaled_axis(glucose_raw)
}
\arguments{
\item{glucose_raw}{dataframe of a valid Libreview glucose table}
}
\value{
scale_x_datetime object, to be added to glucose plot
}
\description{
Adjust x axis depending on time scale of glucose data frame
}
