% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tastermonial-package.R
\docType{package}
\name{tastermonial}
\alias{tastermonial}
\alias{tastermonial-package}
\title{Tastermonial: a Personal Science package web app for CGM analysis}
\description{
A standalone web app that can be run using Docker
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/personalscience/taster}
  \item Report bugs at \url{https://github.com/personalscience/taster/issues}
}

}
\author{
\strong{Maintainer}: Richard Sprague \email{sprague@personalscience.com} (\href{https://orcid.org/0000-0002-3789-5747}{ORCID})

}
\keyword{internal}
