% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\docType{class}
\name{User-class}
\alias{User-class}
\alias{User}
\alias{initialize,User-method}
\alias{show,User-method}
\alias{get_id,User-method}
\title{Set up a new user}
\usage{
\S4method{initialize}{User}(.Object)

\S4method{show}{User}(object)

\S4method{get_id}{User}(x, ...)
}
\value{
user object
}
\description{
Set up a new user
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: initialize with `id = ` and a name asking user to log in

\item \code{show}: Show basic characteristics of object

\item \code{get_id}: User ID
}}

\section{Slots}{

\describe{
\item{\code{id}}{User ID}

\item{\code{name}}{User name}
}}

