% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{user_accounts_update}
\alias{user_accounts_update}
\title{Update User Accounts Databases}
\usage{
user_accounts_update(con, user)
}
\arguments{
\item{con}{valid database connection}

\item{user}{list containing `firebase_id` and `user_id`}
}
\value{
logical TRUE if success
}
\description{
In general, the `user_*()` functions do not have side effects. But sometimes you discover a
brand new `user_id` that doesn't yet map to the database. When that happens, this function is for you.
Given a user list containing a `$firebase_id` and a `$user_id`, it will update the
`accounts_user` and `accounts_firebase` database tables to reflect the new information
}
