% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db.R
\name{user_df_from_db}
\alias{user_df_from_db}
\title{All user records in the database}
\usage{
user_df_from_db(conn_args = config::get("dataconnection"))
}
\arguments{
\item{conn_args}{database connection}
}
\value{
dataframe of all user records
}
\description{
All user records in the database
}
