% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{user_feedback}
\alias{user_feedback}
\title{Feedback to User}
\usage{
user_feedback(output, msg = "Thank You For Listening")
}
\arguments{
\item{output}{output}

\item{msg}{character message to display}
}
\value{
renderText object
}
\description{
A generic way to send feedback to the user.  Assumes a `user_message` tag in the UI.
If the message is NULL, return 'No Value Found'.
}
