% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{user_find_id}
\alias{user_find_id}
\title{Find Unique `user_id`, Creating One If Necessary}
\usage{
user_find_id(con, user)
}
\arguments{
\item{con}{valid database connection}

\item{user}{list containing information needed to set up a new user (either `user_id`, or`firebase_id`)}
}
\value{
list that can uniquely identify the user, including new `user_id` if necessary
}
\description{
Intended to be called after someone has firebase credentials, either
because they are an existing user who is logging in again, or because they are a brand new
user. Either way, this function will return a list that can uniquely identify what
we know about that user.  Besides a unique `user_id`, the list will include items
for the first name, last name, firebase id, and other information.

Ultimately this function should return a user object.
}
\examples{
#con <- db_connection()
me <- list(user_id = 1234, first_name = "Richard",
 last_name = "Sprague",
 firebase_id = "769d1YgcNfTy4rQlxTuMqWR0b3t2")
u = list(first_name = "a",
         last_name = "z",
         user_id = NULL,
         firebase_id = "a1")

#user_find_id(con, u)
}
