% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{user_id_max}
\alias{user_id_max}
\title{Max User ID}
\usage{
user_id_max(con)
}
\arguments{
\item{con}{valid database connection}
}
\description{
Useful for generating a new `user_id`, this returns the current maximum.
}
