% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{user_new_unique_id}
\alias{user_new_unique_id}
\title{Make New Unique `user_id`}
\usage{
user_new_unique_id(con)
}
\arguments{
\item{con}{valid database connection}
}
\value{
integer ID
}
\description{
Each time you call this function, you will get a proposed new and unique `user_id`.
Be careful: this function doesn't save the ID generated, so if you want to be sure that your new ID is
truly unique, you better update the database table.
Obviously, a future version of this function will do that automatically, but right now everyone
who uses this function needs reassurance there are no side effects.
}
