% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{UserObject}
\alias{UserObject}
\title{Create New User Object}
\usage{
UserObject(con, user_id = NULL, firebase_obj = NULL)
}
\arguments{
\item{con}{valid database connection}

\item{user_id}{User ID}

\item{firebase_obj}{Firebase object}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A User object makes for an easy way to
}
\details{
user_management
}
