% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{util_current_user}
\alias{util_current_user}
\title{Current User}
\usage{
util_current_user(user_object)
}
\arguments{
\item{user_object}{User object in list form (e.g. list(con = db_connection(), firebase_id = firebase_setup())}
}
\value{
user_id user ID
}
\description{
Intended to be called as a reactive, this function requires a currently-running firebase instance
and will not respond unless the user is logged in.  If logged in, then return the current user_id
}
