% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_goddess.R
\name{y_scale_}
\alias{y_scale_}
\title{Make Y-Scale Based on All Food Times Values}
\usage{
y_scale_(foods_all, normalize)
}
\arguments{
\item{foods_all}{dataframe of `food_times_df`}

\item{normalize}{logical should return list be normalized}
}
\value{
list
}
\description{
Make Y-Scale Based on All Food Times Values
}
