% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analysis.R
\name{build_all_AUC}
\alias{build_all_AUC}
\title{analysis}
\usage{
build_all_AUC(s_list, cgm_data)
}
\arguments{
\item{s_list}{vector of strings with the names of the foods}

\item{cgm_data}{CgmObject}
}
\value{
The return value, if any, from executing the function.

dataframe
}
\description{
All iAUC values for a vector of food names
}
