% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{user_find_id}
\alias{user_find_id}
\title{Find Unique `user_id`, Creating One If Necessary}
\usage{
user_find_id(con, user)
}
\arguments{
\item{con}{valid database connection}

\item{user}{list containing information needed to set up a new user}
}
\description{
Find Unique `user_id`, Creating One If Necessary
}
