% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user_management.R
\name{UserObject}
\alias{UserObject}
\title{Create New User Object}
\usage{
UserObject(con, user_id = NULL, firebase_obj = NULL)
}
\arguments{
\item{con}{valid database connection}

\item{user_id}{User ID}

\item{firebase_obj}{Firebase object}
}
\value{
The return value, if any, from executing the utility.
}
\description{
A User object represents all the useful information to know about a user:
`user_id` the unique ID used throughout the app. If 0, then the user is unregistered.
`firebase_id` if user is logged in, this is their valid Firebase ID.
You cannot have a `firebase_id` without a non-zero `user_id`
`privilege` char string representation of the privilege level ("admin", "user")
}
\details{
user_management
}
